Fix #3 - Fix broken Flow
Running `yarn flow` gives a weird error, very similar to
[this](https://github.com/verekia/js-stack-walkthrough/issues/5).

This commit resolves the issue, but it seems like an awkward, brittle
overly specific patch. I'm not sure what the more general solution would
be, so I'm going with this fix for now. At least it gets Flow typechecks
up and running.