Add a ping monitor
The app now pings a "deployedUrl", if supplied in config. If the ping
returns anything below a 400, then you get a greent dot.

The idea is to have a minimal "Are you alive?" check, which should give
immediate visibility if a deployment goes down.

This is a very simple and naive test, but hopefully this can form the
seed of something more robust.